refactor: reorganize app structure and introduce new element bindings
- Updated import paths for app elements to reflect new directory structure.
- Added new types for app elements, including inputs, buttons, dialog, and preview elements.
- Implemented binding functions for input handling, style inputs, and dialog controls.
- Created utility functions for reading app state and rendering hexagons.
- Added tests for app element retrieval and state reading functionality.
