+++ src/app/bindRenderInputs.ts
