+++ src/app/bindStyleInputs.ts
