+++ src/app/readAppState.ts
